修复userid过长的问题

+++ MiotSDK/src/main/java/miot.kotlin/model/miot/MiotDevices.kt
@@ -54,7 +54,7 @@ data class MiotDevices(
 
             data class Owner(
                 @SerializedName("nickname") val nickname: String,
-                @SerializedName("userid") val userid: Int
+                @SerializedName("userid") val userid: Long
             )
         }
