+++ app/src/main/java/com/github/miwu/logic/database/model/MiwuDevice.kt
@@ -133,6 +133,6 @@ data class MiwuDevice(
 
     data class Owner(
         val nickname: String,
-        val userid: Int
+        val userid: Long
     )
 }
